Improved testing.

+++ tests/test_meterfiles.sh
@@ -3,8 +3,10 @@
 PROG="$1"
 
 mkdir -p testoutput
-
 TEST=testoutput
+
+TESTNAME="Test that normal meterfiles are written"
+TESTRESULT="ERROR"
 
 rm -f /tmp/MyTapWater
 cat simulations/simulation_c1.txt | grep '^{' | grep 76348799 | tail -n 1 > $TEST/test_expected.txt
@@ -13,9 +15,15 @@ cat /tmp/MyTapWater | sed 's/"timestamp":"....-..-..T..:..:..Z"/"timestamp":"111
 diff $TEST/test_expected.txt $TEST/test_response.txt
 if [ "$?" == "0" ]
 then
-    echo Meterfiles OK
+    echo OK: $TESTNAME
+    TESTRESULT="OK"
     rm /tmp/MyTapWater
 fi
+
+if [ "$TESTRESULT" = "ERROR" ]; then echo ERROR: $TESTNAME; exit 1; fi
+
+TESTNAME="Test that meterfiles with name-id are written"
+TESTRESULT="ERROR"
 
 rm -rf /tmp/testmeters
 mkdir /tmp/testmeters
@@ -25,9 +33,15 @@ cat /tmp/testmeters/MyTapWater-76348799 | sed 's/"timestamp":"....-..-..T..:..:.
 diff $TEST/test_expected.txt $TEST/test_response.txt
 if [ "$?" == "0" ]
 then
-    echo Meterfiles dir OK
+    echo OK: $TESTNAME
+    TESTRESULT="OK"
     rm -rf /tmp/testmeters
 fi
+
+if [ "$TESTRESULT" = "ERROR" ]; then echo ERROR: $TESTNAME; exit 1; fi
+
+TESTNAME="Test that meterfiles with id are written"
+TESTRESULT="ERROR"
 
 rm -rf /tmp/testmeters
 mkdir /tmp/testmeters
@@ -37,6 +51,9 @@ cat /tmp/testmeters/76348799 | sed 's/"timestamp":"....-..-..T..:..:..Z"/"timest
 diff $TEST/test_expected.txt $TEST/test_response.txt
 if [ "$?" == "0" ]
 then
-    echo Meterfiles naming OK
+    echo OK: $TESTNAME
+    TESTRESULT="OK"
     rm -rf /tmp/testmeters
 fi
+
+if [ "$TESTRESULT" = "ERROR" ]; then echo ERROR: $TESTNAME; exit 1; fi
